requirements.txt
streamlit
pandas
supabase
Pillow